+++ src/system/history/sync/files.rs
@@ -409,6 +409,7 @@ mod tests {
                 encrypt: true,
                 variants: vec![],
                 exclude: vec![],
+                include: vec![],
             }],
             ..Default::default()
         };